Component update and implementation of inventory

+++ src/rosecommon/include/components/advancedinfo.h
@@ -10,7 +10,7 @@ struct AdvancedInfo {
 
     AdvancedInfo() : zuly_(0), hp_(0), mp_(0), moveMode_(AdvancedInfo::RUN), runSpeed_(0), atkSpeed_(0), weightRate_(0) {}
 
-    uint32_t zuly_;
+    uint64_t zuly_;
     uint32_t hp_;
     uint32_t mp_;
     uint8_t moveMode_;
